+++ server/spec/factories/interaction_claim_type.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :interaction_claim_type do
-    type { 'ligand' }
-    definition { 'In ligand interactions, a drug forms a complex with its target protein to serve a biological function.' }
+    sequence(:type) { |i| "Interaction claim type #{i}" }
+    definition { 'definition of interaction claim type' }
   end
 end
